refactor: address AI loan pills PR feedback
- Fix util filename typo (aiLoanPIillsUtils -> aiLoanPillsUtils) and update imports
- Consolidate loan fallback into a single safeLoans constant in withAiPills
- Simplify redundant aiPills length check to use the empty fallback directly
- Rename suggestedLoans to randomLoans in EmptyBasketCarousel
- Fill missing @param/@returns JSDoc in the util
